Add service layer
